requirements_email.txt
# Dependências do Celery e Redis
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0
flower>=2.0.0  # Opcional: monitoramento visual de tasks
